+++ R/aaa.R
@@ -1,0 +1,4 @@
+#' @useDynLib lz4lite, .registration=TRUE
+NULL
+
+