+++ NEWS.md
@@ -1,3 +1,8 @@
+# lz4lite 0.2.0 2020-09-24
+
+* Feature: Added `lz4_serialize()` and `lz4_unserialize()` to serialize and
+  compress arbitrary R objects.
+
 # lz4lite 0.1.1
 
 * Added support for vectors of complex numbers